chore: modify folder structure
Please review code changes and approve
