+++ src/javaCourse/homework/lesson2/AnimalType.java
@@ -1,4 +1,4 @@
-package java.course.homework.lesson2;
+package javaCourse.homework.lesson2;
 
 public enum AnimalType {
     CAT,
