+++ src/javaCourse/homework/lesson2/Client.java
@@ -1,4 +1,4 @@
-package java.course.homework.lesson2;
+package javaCourse.homework.lesson2;
 
 import java.util.ArrayList;
 import java.util.List;
