+++ src/javaCourse/homework/lesson2/Dog.java
@@ -1,4 +1,4 @@
-package java.course.homework.lesson2;
+package javaCourse.homework.lesson2;
 
 public class Dog extends Animal {
     public Dog(String name, int age, String breed, String lastVisitDate, String medicalHistory) {
