+++ src/javaCourse/homework/lesson2/Procedurable.java
@@ -1,4 +1,4 @@
-package java.course.homework.lesson2;
+package javaCourse.homework.lesson2;
 
 public interface Procedurable {
     String getName();
